requirements.txt
grpcio-tools
protobuf
websocket-client
redis
pyarrow
requestsweb3
